requirements.txt
django
djangorestframework
environs
djangorestframework-simplejwt
ffmpeg-python